By default, throw if port is inactive when opening
`open_device_by_name` now takes an optional port_num argument.  If it is
given (and not `nothing`) then the port will be queried after device
`dev_name` is opened and an exception will be thrown if the port is not
in the IBV_PORT_ACTIVE state.

The `Context` constructor now takes a `force` keyword argument.  If it
is `false` (the default), then an exception will be thrown if
`device:port_num` is inactive.